Prevented json array of numbers appending on merge
If 2 JSON files are merged into the same JSON parser (by calling Parse(first) then
calling Parse(second) ), then currently, nodes in the tree with the same
key but with differing types are overridden, but object and array
types are merged.

This is problematic for styling, where colors are represented as JSON arrays.

Changed matching arrays of numbers to override rather than merge.

Change-Id: Ia7a37a16919aa8648e7fe34430c5abf04cfaddb4
Signed-off-by: David Steele <david.steele@partner.samsung.com>