Added "mapping" section parser to Builder.
This can be used by any part of the builder to share objects: the
"mappings" section contains key:value pairs. Any value in the parsed
file that contains "<" + key + ">" is replaced by the corresponding
value from the "mappings" section.

Note also that values in the mapping section can also reference other
keys in the mapping section - there is cycle checking, and this will
leave a value empty on the second reference of a cycle.

test-button and test-animation-data test out some concepts that may
make their way into mainline codebase in the near future, such as a
parser to translate property maps into animation data.

Change-Id: I3bf0392c79861fc3433cefc0ad2a3fc610ff58da