Fix issues with text scrolling & transition effects using GetVectorXY() incorrectly
Vector3::GetVectorXY() returns a ref (or a const ref). This is problematic in the files in this
patch because they are returning a ref to a temporary variable which then goes out-of-scope later

The fix is to just create a copy of the value on the stack rather than pass a ref to the temporary
variable.

Change-Id: I2e93748943487a677d394064b70c2b6afe983dbf