Patch coverage tool
After building and running automated tests, you can test
the coverage of your current patch (either in the working
directory, or the head patch, or some other difference to HEAD
supplied by the arguments).

The script will return 1 if the test coverage is less
than 90%, otherwise it will return 0 to indicate no error.

This script will also be run as part of the verification process.

Change-Id: Id931d37ab0bdc8dcb179d34eed69d093efd7286f
Signed-off-by: David Steele <david.steele@samsung.com>