[3.0] Updated patch-coverage tool to catch fewer errors
Patch coverage tool was catching coverage errors for files in
non-source tree and of non-source code, also was catching changes
that had no coverage change.

Also added mechanism to choose the right patch, or to fail
if there are changes in both the working tree and the index.

Also added html output.

One minor gotcha spotted - added test code which was in-advertently covered
because of a case fallthrough, and passed, though this then caused unchanged
code to reduce in coverage.

Change-Id: Id49d30935b2dc9ef3145bf6bc78ec7510de01fd0
Signed-off-by: David Steele <david.steele@samsung.com>