Added code for stylable transitions
This change enables animatable properties of placement actors/renderers
to be animated using a property map defined by a style sheet.

New classes:
VisualProperty - a version of Property that doesn't use C++ references
TransitionData - a pimpled struct that carries enough information to generate
Animators that can be defined by Animation::AnimateTo().

Additions:
Added a name to visual base to enable the animator to target a named
visual.

Added GetVisualProperty() internal method which checks the named visual
(or it's placement actor) for the given property, and returns the appropriate
VisualProperty that can animate it.

Change-Id: I0c91eae845b0ba5be5d7829babee75717f2b3b49
Signed-off-by: David Steele <david.steele@samsung.com>