DALi C# binding - adding property for Animation LoopCount and others.
Implemented setters and getters for the property LoopCount in Animation Class
and the X,Y,W and H properties in RectInteger Class. Combined the opacity and
CurrentOpacity properties in Actor class. The properties Size(from Stage class)
and Visibility(from Actor class) are already present.

The Patch also contain the Changed order of the types in the generic parameters to
match the declaration in AnimateBy and AnimateTo APIs.

Change-Id: I8286f4ae19c40b3911e0d12f6874f7c2b184c26c
Signed-off-by: Ravi Nanjundappa <nravi.n@samsung.com>