Dali C# binding : Creating Color Constructor to accept enum Colors
The Color class takes a string in the constructor, so it cannot check
whether the string represents a valid color. Here we define constants for
these colors as enum, so that it can do this check during compilation.

Now we can use stage.BackgroundColor =  new Color(Colors.White); instead of
stage.BackgroundColor =  new Color("white");

Change-Id: I135e4103d6f9171011ed9d4e60f486a3e237d6cd
Signed-off-by: Emil Abraham <emil.abraham@samsung.com>