Updated control transitions to work with visual transform
Modified Transform::SetPropertyMap() to handle string keys
( for styled transitions that set initial values )

Added GetRenderer to Internal::Visual::Base (not public)
( To allow control to access renderer's registered properties )

Changed Control::Impl::GetVisualProperty to ignore actor properties
and instead only use renderer properties. ( Allows transform::size
to be accessed )

Ensured RegisterVisual automatically names un-named visuals to
match the name of the property they are associated with. (This
makes it easy to wire up styled transitions to visuals).

Added a new method AddTransitions() to Internal::Control. This
enables the styled transition framework to add multiple animators
to a single animation. (slightly controversial - this is a new
public API, which should be a Devel API, but is really hard to
implement - an alternative approach is currently being looked at,
on which this may have to be rebased.

Change-Id: Ia3a1365594a1367fec9af09b9481f8696aef1d76
Signed-off-by: David Steele <david.steele@samsung.com>