Published the AsyncImageLoader and SyncImageLoader to the public API
The AsyncImageLoader API allows asynchronous image loading via an internal thread.
Multiple images can be loaded with one AsyncImageLoader.
The load of each image returns an ID.
This ID is returned, along with the pixel data in a signal once the load for each individual image has completed.

The SyncImageLoader allows synchronous image loading.
It consists of a set of Load functions mirroring the prototypes of the Load functions within the AsyncImageLoader.
The loads are performed within the same thread (the thread is blocked whilst loading).

Change-Id: If1d59bf5981de5f52371bd5838371b41031d1fd4