Updated visuals to add animatable mixColor property
Modified all visuals to add a mix color property, except
color visual, which already has one, and primitive visual,
which has such a property, but have moved it to register on
the renderer rather than the shader.

Color and Primitive visual will handle either MIX_COLOR key, using
their own key in preference to the base key.

The target mix color and property index is now held in the
Visual::Base impl class. The uniform registration is also done
in the base clas.

Added handling to control to enable setting or animating the
mix color through the style sheet.

Added a note to Internal::Control::CreateTransition() to warn that
no animation will be created if the control is not staged or the
visual is not enabled

Ensured the blend mode is set to the right value for the target
mix color when the transition finishes.

Change-Id: I3191da06098582f57f101a8a0139faa3b1fb3c80
Signed-off-by: David Steele <david.steele@samsung.com>