Reduced size of control depth index ranges
The control depth index ranges, used for setting the depth index of
visuals, needs to stay within the sibling order range, which is
defined as 1000. The existing values for background (used by button)
overlapped with sibling order and caused incorrect ordering in the svg
demo.

Of note is the Layer::TREE_DEPTH_MULTIPLIER is different from the
DevelLayer::TREE_DEPTH_MULTIPLIER value, and should be removed.

Change-Id: I176f277286840f5b57840661f285dc9261795769
Signed-off-by: David Steele <david.steele@samsung.com>