Updated ImageVisual to handle CPU Image Masking
Updated ImageVisual to add a URL for an alpha mask image. If this is present
in the initialization map, then the mask is loaded and stored in CPU (it
may be used by other textures). On triggering the Async image loading in
TextureManager, the alpha mask's textureId can be passed in.

Added new state machine to TextureManager to handle loading a mask and
and an image; and to apply the mask to the image when both have loaded.

Image masks are applied in CPU side, so they are stored in CPU memory.
All other images are uploaded to GPU ( after optional mask has been applied )

Change-Id: I6cce7f62d6d6765dc6199bb25891791333533dea