(C#) Changed Rectangle to have float properties
Wraps Rect<int> in C++. Problem with ints is that an explicit conversion needs to be done in C#
application code which looks bad. Int to float conversion is automatic so will still work.

Made a comment in the property to state that the decimal is ignored and that the float is provided
for convenience only.

Change-Id: I9589999406c01529c7f2463c07930e7d1d1bdca6