Added scrollMode property to ScrollView to simplify Rulers
The rulers are hard to use and understand, and require overriding
classes in the public API. This makes binding to C# require too much
effort, and pulls too many classes into C# that otherwise don't need to
be there.

Have instead added a devel property to enable the simplification of
setting up rulers - the new property map will be used internally to
create either DefaultRuler or FixedRuler, depending on the
settings. This makes it very easy to bind to C#.

Change-Id: I5932cd32487bae77b8c41ebd35e35f6bc879be33