Adding support for multiple images in AnimatedImageVisual
As well as supporting animated GIF, the animated image visual also
supports animating between a number of images.

This can be done using a property map containing the following:

{
   "url":[ "url1", "url2", "url3", "url4", "url5" ],
   "batchSize" : 3,
   "cacheSize" : 4,
   "frameDelay" : 100
}

The cache size is the total number of images that will be loaded at any one time;
the batch size is the maximum number of images it will attempt to load before displaying.
If the cache size is smaller than the batch size, it is increased to match. If either
are larger than the number of urls, they are reduced to match.

As soon as the visual is created, it will start loading up to batchSize urls into the
cache. After each frame has completed, it will start loading the next batch, etc,
until the cache is full. After the next frame, it will load only the next URL into
the cache.

If at any time, the next image is not ready, it will wait for the loading to complete
before displaying and triggering the next batch/frame load, but will not reset the timer.

Change-Id: Iccd59768aec814ab3b73fb7b817e45b95299b2c2
Signed-off-by: David Steele <david.steele@samsung.com>