Adding Layouting classes to Toolkit
Adding new layouting engine based on Android. Classes LayoutBase, LayoutParent and LayoutGroup
mirror View, ViewParent and ViewGroup from Android and are responsible for measuring and
layouting. Actual layouters such as TableLayout and FlexLayout will inherit from LayoutGroup.

Layout metadata for controls and visuals that are being laid out is stored in ChildLayoutData and
derivatives. MeasureSpecs are used by layouters to define how their children react to measuring.

Layouting is performed at the end of the event frame by the LayoutController.

Change-Id: Iebf82323be8d432fb8e1f815d83b56b1d1a1cde4
Signed-off-by: David Steele <david.steele@samsung.com>