ImageVisual Load policy
ImageVisual can be created with a PropertyMap that inserts a load policy.

Depending on the load policy the image(texture) will load immediately when the visual
is created or when the visual is attached to the stage.

Can be used to preload special images so can be displayed immediately on request.

Registered Visuals with LoadPolicy::IMMEDIATE will signal ResourceReadySignal when attached to stage.

Limitations:

Atlasing not supported with IMMEDIATE mode currently.

Change-Id: I3e1b3fd19f3027edb62858549890972f6a166cd0