ImageVisual API for Orientation Correction
Added Property OrientationCorrection which can control if the image should be rotated
to it's Orientation as defined by exif data.

Provided for JPEG images containing the exif data for Orientation.

Limitations:

The image provided will be cached, the orientation of the first cached image will be
used for subsequent images until the cache image is released.

Change-Id: I7627839c143c06dfde63a32eb8171e2205bf18af