Action framework for visuals
Internal::Visual::Base has a DoAction API which calls the OnDoAction in derived Visual if they want
 to act on any Actions they have defined.

Control has a devel API to DoAction on any of it's registered visuals.

If a action is requested on a visual, if the visual does not implement it then
 nothing is done.

Action will be provided by Indexes,  in future may add an API for strings if
 builder requires but internal.
 Property::Key has an explict constructor hence not used.

DummyVisual added to test framework, can be used in Internal tests. Visuals creation not public.

Change-Id: Ib4590f5a76d4b45adcb2213eee70c1bbd91f88a2