Removed need for a layout container view
Deleted HboxView pimpled class
Removed the need for a Layout to be owned by a specific container type.
Instead the layout can be attached to any control. It automatically
listens for controls being added to / removed from the control, and adds them
to itself. (It does not currently pick up existing children - suggest we add a
new API for this)

(There is currently no specific handling for visuals).

Change-Id: Icf8326df4b351093810352dbc2452505ccb5879d
Signed-off-by: David Steele <david.steele@samsung.com>