Fixed unparenting a layout item from a layout group
Unparenting a layout item from a layout group did not actually remove the
child from the parent, which if a new layout were set on a container, resulted
in the container having 2 separate layouts in it's parent layout.

Changed the Layout Parent/Child interfaces to allow LayoutItem
to remove itself from a LayoutGroup without explicitly referencing it's subclass.

Change-Id: Idf6d7813ab5a440dfbf44a8bbe46df2ca3698d64
Signed-off-by: David Steele <david.steele@samsung.com>