Added base64 encoding/decoding for Property Array/String
We need a means of reading and writing data blocks through the property
system for CustomShaders in VK. Added a mechanism to allow encoding
a vector of uin32_t data into a Property::Value of type STRING or
ARRAY of STRINGS, and a means of decoding the same.

This allows us to keep the property system in line with a JSON encoding,
but trading off data size and decode time.

Change-Id: I2247b340571644b21be5b9ef87fc0e3108e2f571
Signed-off-by: David Steele <david.steele@samsung.com>