BinLayout added
This layout has children but does not explictly position them itself.
If given a (legacy) container as a child it will check it for ResizePolicies
 and map them to a MeasureSpec so can be used by the Layout system.

The BinLayout by default will take the size of the greatest child dimensions.
(WRAP_CONTENT).
Although can be set to MATCH_PARENT.

PARENT_ORIGIN as always TOP_LEFT, may account for this in future if required.

Change-Id: I2335ee465578d796e1025192ae0f8cdb95ecc60f