Stripped down GaussianBlur view to only blur a scene.
Removed functionality to blur a single image (this should be a separate effect view)
Removed API to choose a pixel format - Vulkan supports attachments with RGBA8888
Kept API to alter number of samples, but this is ignored - it should be re-added when
we have Specialization Constants in shader code.
Removed ActivateOnce and FinishedSignal - currently can only support REFRESH_ALWAYS
as we don't know when all of scene is ready. Application is better placed to manage
resource ready handling.

Change-Id: Id0e1cde7d260f5ba2b9edb958aebe82a962521cf
Signed-off-by: David Steele <david.steele@samsung.com>