[Tizen] Store LayoutItem target size when set explictly
MinimumWidth and height were used to store the set size of LayoutItems,
this was a problem if the ResizePolicy changed and the size set was no longer
needed as it was still stored.

If the resize policy size was less than the minimum size stored then it would
not be updated.  Hence Actor would not re-draw it after SetFrame called.

TargetSize has been added to the LayoutItem now.  It can be set to zero when no
longer needed and we can still have Minimum size working.

Change-Id: Id3f0cec1d4b1e991f2dd7034c2895ed7d8decd6c