Change to output in priority order.
ex)
The current character(0x2665) is emoji script.
'UIFont' and 'EMOJI' font have '0x2665' unicode.
The priority of font is 'UIFont' -> 'EMOJI'.
The now behavior is that we find and output glyphs in EMOJI fonts. because the current character is an emoji script.
However, we want to be output from a high priority 'UIFont'.
so modified it.

and If you want to print emoji fonts, you can add variation unicode.
0x2665+variation unicode = at 'EMOJI' font file
The logic to check variation unicode will be added later.

Change-Id: Id5b54a75280d1678874665a42291facdbce72042