Fix text jitter issue
Text jitter issue occurs when set a text again during text scrolling.
It is because the same property is used by the old and new animations. 2 PropertyResetters affect the same property.
So I've changed to use properties of Shader which is newly created every time.
Then the same property is not used by the old and new animation now.

Change-Id: I7ec9b266b0be48414d8df5e3ffead6e5c750b338