[Tizen] Add VisualBase::GetType()
Some controls want to know the visual type when the visual is given from the user.
To do that in current DALi, it needs to call CreatePropertyMap() and get TYPE property from the map.
Instead, this patch provides an API Visual::Base::GetType() to make it easy to see what the visual type is.

Change-Id: I387a6cc43466897b2de17d3c2d9bae7dba201a10
Signed-off-by: Jiyun Yang <ji.yang@samsung.com>