Call LoadTexture() for ReleasePolicy
When ReleasePolicy is not DETACHED, There are cases where the wrong
texture is seen.
To avoid this case, i added a patch.
Detailed description is added below.

Load Texture if mTextures is empty.
mTextures is already set, the mTexture can be used to create Renderer.
There are two cases mTextures is empty.
1. mTextureId == TextureManager::INVALID_TEXTURE_ID and mTextures is empty
   - Visual is on stage with LoadPolicy::ATTACHED
   - mTextureId != TextureManager::INVALID_TEXTURE_ID and mTextures is empty
   - If ReleasePolicy is DESTROYED, InitializeRenderer called every on stage called.
   - Then every resources those contained in Visual are Reset but mTextureId is remained when the Off stage time,
   - So, mTextures needed to be get from texture manager to created resources like mImpl->mRenderer.

Change-Id: Ice111ee467eaeba222eaa9063bde3b23ada484d4