[windows] Use unsigned for enums used as bitfields
When using enums as bitfields, the Microsoft C++ compiler will generated
shift intructions to when comparing the enum with one of its constants.
As by default enums are signed interegers, the compiler will generated
signed shifts, yielding trash values.

We define the enums in the TextureInfo class as unsigned ints to avoid
this bug.

Change-Id: Icdd49c3b417b09313964c71381b7c8230c844f12