Ensure AnimatedVectorImageVisual doesn't hang app
Currently, if the AnimatedVectorImageVisual is not destroyed
on Application::Terminate, e.g. if it's actor is kept alive,
then the shutdown order is undefined, and may end up
trying to pend on a mutex which has already been destroyed,
in the dali-extension plugin.

(TizenVectorAnimationManager, which creates the mutex, is a
singleton that is created when first needed, but stays alive
until main() completes)

Prevented the destructor from finalizing in the plugin if
core has already been shutdown.

Change-Id: I9fe547eab4dc3f3e5a4655f5c8dcf96bf664d7e3