Avoid shader string copy when possible.
As shader creation will take a copy if the shader is not in the cache
we don't have to create new String for each shader creation request.

we create a new string only when we need to update the shader data
and only pass a string_view to the shader creation.

Change-Id: I595d5f2af8216c2526ba0865fe33b6c0c2538405