[Tizen] fix wrong cursor position in TextField
When the placeholder font size is bigger/smaller than default font size,
the cursor is positioned incorrectly.

this is because mScrollPosition.y is calculated based on layoutSize.height at CalculateVerticalOffset().
(this is not based on cursor height.
in issue situation, layoutSize.height is bigger/smaller than defaultFontLineHeight.)

This patch prevents the wrong positioning of cursor when the layoutSize.height is bigger/smaller than defaultFontLineHeight.
And update the glyphPositions using recalculated offset.

This issue only occurs with placeholder + single line text control. (vertical alignment)

I hope this modification does not affect the calculation of other texts.
So, update the glyphPositions in CalculateVerticalOffset().

Change-Id: I05847bbe5dcce6451c6b24649f80ea945e1ed7ea
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>