Fix ClearHighlight() for custom highlight frames
ClearHighlight() used GetHighlightActor() to retrieve the currently
displayed highlight frame in order to determine which actor to remove
from the scene graph. However, it is perfectly legal to
SetHighlightActor() while a highlight frame is displayed (takes effect
on the next GrabHighlight()), so the currently visible highlight frame
handle needs to be saved separately in order to be properly cleared.

Change-Id: Ibf41574bac66f9375b72b31d061195d0235621aa