[Tizen] 1. Implement FocusFinder
The class used for finding the next focusable actor in a given direction
from a actor that currently has focus.

2. Change FocusFinder to a method

3. Add Tab key focus movement.

If the focus group is not changed, Move the focus towards right, "Shift-Tap" key moves the focus towards left.
Still if the focus is not moved, Move the focus towards down, "Shift-Tap" key moves the focus towards up.

4. Users can implement focus movement by SetCustomAlgorithm.
So User implemented focus movement by SetCustomAlgorithm(AutoFocusAlgorithm.Instance);
And if user do SetCustomAlgorithm(null), the user expects the focus not to move.
this is for backward-compatibility.
Add EnableDefaultAlgorithm() and IsDefaultAlgorithmEnabled() for enable/disable the default algorithm

5. Fixed bug

6. Fixed IsCandidate bug

7. Check if actor can be focused.

8. If parent is hidden, the child cannot be focused.

Change-Id: Ieebaafc31fe5e943a63b8e2a079335eb578ab0cd