[AT-SPI] Call Control::Impl::AccessibilityRegister() unconditionally
This patch fixes a bug where the screen reader is not notified about
changes of size / position / visibility of controls if accessibility is
enabled in runtime, i.e. after the controls are connected to the scene.

Please note that the callbacks registered with AccessibilityRegister()
already check Accessibility::IsUp(), therefore non-accessibility usage
scenarios are not affected by this patch.

Change-Id: Icc49b0a12a75135ca3d2033b342878b4ab1392bb