[AT-SPI] Do not emit BoundsChanged for all actors
This patch also removes the ACCESSIBILITY_ANIMATED property (not to be
confused with AT-SPI state ANIMATED). This property was used to control
automatic emission of the BoundsChanged event (and it had only one user,
the highlight frame actor). However, we no longer emit this event, so
this property serves no purpose.

Change-Id: I8df8b9e7fb75951c9f88d5aec944f9d52f10bf06