[AT-SPI] enhance SHOWING state decision logic
To define SHOWING state, using CULLED and VISIBLE property is not enough.
If a child is clipped out by its parent, then CULLED property is not TRUE.
We need to use similar logic as EFL does.

./src/lib/elementary/efl_ui_widget.c > _elm_widget_onscreen_is
is used for calculating SHOWING state using parents geometry information.

So this patch set is using parents extent information for SHOWING state.

Change-Id: I5eecf93c66ce199bd0ca88f3ad3ff471c1ff79a5