[AT-SPI] Set ImageView non-highlightable by default
Removed the line setting ACCESSIBILITY_HIGHLIGHTABLE property to true.

Until now the highlightability of the ImageView was overrided by an
AT-SPI Role check in dali-adaptor, so the line removed has been already
a dead code. Now the Role check has been removed and we want ImageView
to be non-highlightable by default.

From now on if we want to turn on highlightability for certain
instances of ImageView we can set the ACCESSIBILITY_HIGHLIGHTABLE
property in runtime.

Change-Id: Ide151a114d970f856217b34945d4c7e9cbbecda0