[AT-SPI] Set CanvasView non-highlightable by default
Removed the line setting ACCESSIBILITY_HIGHLIGHTABLE property to true.

Until now the highlightability of the CanvasView was overrided by an
AT-SPI IMAGE Role check in dali-adaptor, so the line removed has been
already a dead code. Now the Role check has been removed and we want
CanvasView to be non-highlightable by default.

From now on if we want to turn on highlightability for certain
instances of CanvasView we can set the ACCESSIBILITY_HIGHLIGHTABLE
property in runtime.

Change-Id: Ie6d695c6aa61f77e34029481e46ccd3b7fea7ac0