Fix SetPrimaryCursorPosition() behaviour when no focus
If TextField does SetPrimaryCursorPosition() before CONNECTED_TO_SCENE,
there is an issue that the caret renders regardless of focus.

Also, SetPrimaryCursorPosition() always sets focus to the Text Control regardless of the user's intent.
This creates unintended bugs in the app.

This patch only updates the value of PrimaryCursorPosition when there is no focus.
When there is focus, the behaviour is the same.

Change-Id: I2bc5851c36e2c08976e7ca4656bfe38f6afe0653
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>