Add TextSelect()
This api can replace the SELECTED_TEXT_START and SELECTED_TEXT_END properties.
The feature to select a range of text should be provided as a method like SelectWholeText(), SelectNone().

Currently, the SELECTED_TEXT_START and SELECTED_TEXT_END properties are causing several problems.
1. User can not set start and end at the same time.
2. Since it is not an event method, it does not work if it is called before text is rendered.

This patch adds SelectText().
And I will make a separate patch to remove the property.

Change-Id: I5a050cd88b9597b939766e1a6ec0e39edf96e2d4
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>