[Tizen] Change raw pointer to unique_ptr
The mPlaceholderFont pointer is allocated memory once only
when the text-controller uses a placeholder.
This will persist until mEventData is destroyed.
Changed to a smart pointer suitable for maintenance.

Additionally,
In SetPlaceholderTextFontWidth() and SetPlaceholderTextFontSlant(), mPlaceholderFont can not be NULL.

Even if there is no value in SetFontStyleProperty, it is always set to None,
This is because an instance of mPlaceholderFont is created through SetPlaceholderTextFontWeight(), which is called first.

So this causes coverage issues.
Replace the duplicate code with CreatePlaceholderFont() solves this problem.

Change-Id: Ibd11b02f22e505be8643c9360b5d5e5e2de1b318
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>