[AT-SPI] Change the type of Control::Impl::mAccessibilityRelations
This patch changes the relation's target type from
`Accessibility::Address` to `Accessibility::Accessible*`.
`AppendAccessibilityRelation()` does not allow to specify a remote
object anyway, because the parameter types are `Actor`s.

Moreover, the `Address` may be incomplete (specifically the
DBus-dependent part) if the above API is called too early. With this
change, the `Accessible::GetAddress()` call is postponed until the
relations are queried over DBus.

Change-Id: I881d7fb90eba79b6ba9752e913d113272485e29a