[AT-SPI] Follow-up on mAccessibilityRelations
The previous implementation used an `std::vector` (the outer one) as a
map, with the `RelationType` key working as the index to that vector.
However, relations are rare in practice, and every `Control::Impl`
having a 23-element vector or vectors, most, if not all of which are
empty, feels like a waste of memory.

This patch also changes the inner vector to a set, which resolves the
previously unaddressed issue of duplicate elements.

Change-Id: I639082fe49e0f6107213cda5b142a09882a3be39