DoAction can chage CornerRadius/Borderline/BlurRadius
we can update properies by DoAction(DevelVisual::Action::UPDATE_PROPERTY).
It is not an usual case, but we may need to prepare of NUI code fix.
(There is some related issue at NUI ImageView)

Here, I update some properies that need to update shader.
We don't want to change the shader many times.
(ex : DoAction with CornerRadius = 1.0f, and DoAction 0.0f after that, and repeat.)
So if someone update properies by DoAction API, turn on
mAlwaysUseXXX = true so shader will not be previous shader anymore.
(This is same logic when these properies are animated.)

Change-Id: I963930f261982d50bd647e4ba7ed4e476dd8898d
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>