[AT-SPI] Change DevelControl functions' first parameter to Control
These functions did nothing for non-Control actors anyway. It is better
to narrow down the parameter type to a Control, especially since they
are in the DevelControl namespace. As a special exception,
SetAccessibilityConstructor() retains its Actor parameter, because
changing that would require too many changes, and we intend to replace
it with a virtual method very soon anyway.

Change-Id: I0b5939dd9fc0933841cd36b6833aeff668d2ab30