Do not create visual when URL or array is empty
When we try to create ImageVisual with empty string, it will be failed.
But when we try to create AnimatedImageVisual with empty string,
it will try to create new AnimatedImageVisual.

AnimatedImageVisual is weak when mFrameCount is zero.
So we try to guard that case before creation-level.

Moreover, we also don't create visual for SVG, N_PATCH, ANIMATED_VECTOR_IMAGE

Change-Id: Ibd7a3db825473958bbe51e6fb909572464b21da6
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>