[AT-SPI] Remove SetAccessibilityConstructor()
This function is replaced by a new Control virtual method,
CreateAccessibleObject(), which allows to achieve the same goal using
less eccentric syntax. A slight reduction in memory usage is also
expected, thanks to the fact that there is now no need to store a
'mAccessibilityConstructor' functor in every Control::Impl instance.

Change-Id: I6399a6ca9acefb1648086b17d154541535f6bd2d