Fix cursor visible issue on text editing
I made a prev patch to ensure that the cursor layer is on top.
16ef097a0b2d3945a5045db7bb2f3896ed5a5a95

But cursor layer in clippingDecorationActors(in RednerText) is processed only once.

Therefore, if user edit while the background is present, the cursor still invisible.

This patch passes the cursor layer actor as an argument to RenderText(),
so that the cursor is always on top.

Change-Id: I27e3334f74dbca34aaae18089082125c871935fa
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>