Block the changeness of PreMultiplied for some visuals
Svg, AnimatedVector, Text, and Gradient Visuals always use
PreMultiplied texture.
But PreMultiplied option is Visual::Base property.

So it was possible to change premultiplied option as false
even It's visual type is not allow it

This patch make some kind of visual that should not change the
PreMultiplied alpha flags don't change it.

Change-Id: I22e8416796493b34d2907f8043f57d31039be587
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>