[AT-SPI] ScrollView: Use child postion for ScrollToChild
It was not able to scroll in a child of scroll view if the child is
not a child of the first depth.

(1) [scroll pane],[(class=ScrollView)],[364,199,656,36],[TextSelectionScrollView]
  (2) [table],[(class=TableView)],[364,101,901,134],[]
    (3) [push button],[(class=PushButton)],[364,101,447,134],[Select all]
      (4) [image],[(class=ImageView)(highlight=)],[364,101,447,134],[]
    (5) [unknown],[(class=Control)],[811,107,1,122],[]
    (6) [push button],[(class=PushButton)],[812,101,265,134],[Copy]
    (7) [unknown],[(class=Control)],[1077,107,1,122],[]
    (8) [push button],[(class=PushButton)],[1078,101,187,134],[Cut]

The node (8) was not able to bring in even though it has highlight.
This patch makes node (8) bring in when it grabs highlight.

Change-Id: I862a8adcc2bd6212464ffc833a2a0d0cf0a4480c